requirements.txt
alembic==0.9.9
aniso8601==3.0.2
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
awscli==1.15.14
botocore==1.10.14
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
constantly==15.1.0
cryptography==2.3
cryptography-vectors==2.3
cssselect==1.0.3
docutils==0.14
Flask==1.0.2
Flask-Migrate==2.2.1
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.2.0
parsel==1.5.0
pluggy==0.7.1
psycopg2==2.7.5
py==1.5.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
PySocks==1.6.8
pytest==3.7.1
pytest-runner==4.2
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
PyYAML==3.12
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
Scrapy==1.5.1
scrapy-redis==0.6.8
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.10
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.5.0
